Make the web-ext config an ES module
web-ext loads its config file with `import()`. On newer Node versions,
importing the `.cjs` config also exposes a `module.exports` named export,
which web-ext rejects: `The config option "module.exports" must be
specified in camel case`. That breaks `npm run build:firefox` before it
can produce an .xpi.

Renaming the config to `.mjs` with `export default` avoids the extra
named export; project.config.ts is still loaded through the sucrase-node
hook via `createRequire`. Update the eslint override, the tsconfig
exclude and the source-bundle file list for the new name.
