+++ .eslintrc.js
@@ -240,7 +240,7 @@ module.exports = {
       },
     },
     {
-      files: ["rollup.config.js"],
+      files: ["rollup.config.js", "web-ext-config.mjs"],
       parserOptions: {
         sourceType: "module",
       },
