+++ scripts/build-post.ts
@@ -58,7 +58,7 @@ async function makeSourceCodeBundle(): Promise<void> {
     "project.config.ts",
     "rollup.config.js",
     "tsconfig.json",
-    "web-ext-config.cjs",
+    "web-ext-config.mjs",
   ].map((file) => path.join(BASE_DIR, file));
 
   const dirs = ["@types", "docs", "patches", "scripts", "src"];
